Move GrainId and Interner to Abstractions
Removed SafeTimer from the Interner copy and just used a regular timer.
